+++ src/page.h
@@ -25,7 +25,7 @@ typedef struct {
 
 void *redislite_page_get(void *_db, void *_cs, int num, char type);
 int redislite_page_register_type(void *db, redislite_page_type *type);
-void *redislite_page_get_by_keyname(void *_db, void *_cs, char *key_name, int length, char *type);
+void *redislite_page_get_by_keyname(void *_db, void *_cs, char *key_name, size_t length, char *type);
 redislite_page_type *redislite_page_get_type(void *db, char identifier);
 int redislite_page_delete(void *_cs, int num, char type);
 #endif
